+++ Sources/Sentry/SentryHub.m
@@ -240,8 +240,10 @@ SentryHub ()
                        withScope:(SentryScope *)scope
          additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems
 {
-    if (transaction.trace.context.sampled != kSentrySampleDecisionYes)
+    if (transaction.trace.context.sampled == kSentrySampleDecisionYes) {
         return SentryId.empty;
+    }
+
     return [self captureEvent:transaction
                       withScope:scope
         additionalEnvelopeItems:additionalEnvelopeItems];