+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - feat: Add flag for UIViewControllerTracking (#1711)
+
 ## 7.12.0-beta.0
 
 ### Various fixes & improvements
@@ -11,7 +12,6 @@
 
 ## 7.11.0
 
-- feat: Add support for profiling on iOS (#1652)
 - feat: Add CoreData performance tracking (#1682)
 - fix: Detecting ANRs as false OOMs (#1695)