+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix: Setting invalid sample range (#1729)
 - feat: Add more info to touch event breadcrumbs (#1724)
 
 ### Important notice